@@ -43,4 +43,4 @@ class Notification {
 
 }
 
-module.exports = Notification;
+export default Notification;
